refactor: share ANSI stripping
