+++ crates/luban_backend/src/services/claude_cli.rs
@@ -14,6 +14,7 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Duration;
 
+use super::ansi::strip_ansi_control_sequences;
 use super::stream_json::{
     extract_content_array, extract_string_field, parse_tool_result_content, tool_name_key,
     value_as_string,
@@ -24,30 +25,6 @@ pub(super) struct ClaudeTurnParams {
     pub(super) worktree_path: PathBuf,
     pub(super) prompt: String,
     pub(super) add_dirs: Vec<PathBuf>,
-}
-
-fn strip_ansi_control_sequences(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    let mut chars = input.chars().peekable();
-    while let Some(ch) = chars.next() {
-        if ch != '\u{1b}' {
-            out.push(ch);
-            continue;
-        }
-
-        if matches!(chars.peek(), Some('[')) {
-            let _ = chars.next();
-            for next in chars.by_ref() {
-                if next.is_ascii_alphabetic() {
-                    break;
-                }
-            }
-            continue;
-        }
-
-        let _ = chars.next();
-    }
-    out
 }
 
 fn resolve_claude_exec() -> PathBuf {
